# ms-solicitudes/src/main/resources/openapi/ms-solicitudes.yaml
openapi: 3.0.3
info:
  title: ms-solicitudes API
  description: Gestión del ciclo de vida de solicitudes de crédito
  version: 1.0.0

servers:
  - url: http://localhost:8080/api/v1
    description: Local vía API Gateway

# ─── Seguridad global ────────────────────────────────────────────────────────
security:
  - bearerAuth: []

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  # ─── Schemas reutilizables ─────────────────────────────────────────────────
  schemas:

    EstadoSolicitud:
      type: string
      enum: [PENDIENTE, EN_REVISION, APROBADA, RECHAZADA]
      description: >
        PENDIENTE    → recién creada, esperando revisión.
        EN_REVISION  → un analista la tomó.
        APROBADA     → crédito aprobado.
        RECHAZADA    → denegada con motivo obligatorio.

    CrearSolicitudRequest:
      type: object
      required: [monto, plazoMeses, proposito]
      properties:
        monto:
          type: number
          format: double
          minimum: 1000
          maximum: 500000
          example: 25000.00
          description: Monto solicitado en soles
        plazoMeses:
          type: integer
          minimum: 6
          maximum: 60
          example: 24
          description: Plazo de devolución en meses
        proposito:
          type: string
          minLength: 10
          maxLength: 500
          example: "Compra de equipos para negocio"
          description: Descripción del destino del crédito

    RechazarSolicitudRequest:
      type: object
      required: [motivo]
      properties:
        motivo:
          type: string
          minLength: 10
          maxLength: 1000
          example: "Historial crediticio insuficiente según bureau"

    SolicitudResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1001
        usuarioId:
          type: integer
          format: int64
          example: 42
        monto:
          type: number
          format: double
          example: 25000.00
        plazoMeses:
          type: integer
          example: 24
        proposito:
          type: string
          example: "Compra de equipos para negocio"
        estado:
          $ref: '#/components/schemas/EstadoSolicitud'
        motivoRechazo:
          type: string
          nullable: true
          example: null
        fechaCreacion:
          type: string
          format: date-time
          example: "2025-06-01T10:30:00Z"
        fechaActualizacion:
          type: string
          format: date-time
          nullable: true
          example: null

    SolicitudPageResponse:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/SolicitudResponse'
        page:
          type: integer
          example: 0
        size:
          type: integer
          example: 10
        totalElements:
          type: integer
          example: 47
        totalPages:
          type: integer
          example: 5

    MetricasResponse:
      type: object
      properties:
        totalPendientes:
          type: integer
          example: 12
        totalEnRevision:
          type: integer
          example: 5
        totalAprobadas:
          type: integer
          example: 89
        totalRechazadas:
          type: integer
          example: 23
        montoTotalAprobado:
          type: number
          format: double
          example: 2250000.00

    ErrorResponse:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
          example: 400
        error:
          type: string
          example: "BAD_REQUEST"
        message:
          type: string
          example: "El monto no puede superar S/. 500,000"
        path:
          type: string
          example: "/api/v1/solicitudes"

# ─── Endpoints ────────────────────────────────────────────────────────────────
paths:

  /solicitudes:
    post:
      summary: Crear solicitud de crédito
      description: >
        Crea una nueva solicitud en estado PENDIENTE.
        El usuarioId se extrae del header X-User-Id inyectado por el Gateway
        — el cliente no lo envía.
        Publica evento solicitud.creada en RabbitMQ al completar.
      tags: [Solicitudes - Usuario]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CrearSolicitudRequest'
      responses:
        '201':
          description: Solicitud creada exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SolicitudResponse'
        '400':
          description: Datos de entrada inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Token JWT ausente o inválido

    get:
      summary: Listar mis solicitudes
      description: >
        Retorna las solicitudes del usuario autenticado (extraído de X-User-Id).
        Soporta paginación y filtro por estado.
      tags: [Solicitudes - Usuario]
      parameters:
        - name: estado
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/EstadoSolicitud'
        - name: page
          in: query
          required: false
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          required: false
          schema:
            type: integer
            default: 10
      responses:
        '200':
          description: Lista paginada de solicitudes
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SolicitudPageResponse'
        '401':
          description: No autenticado

  /solicitudes/{id}:
    get:
      summary: Obtener detalle de una solicitud
      description: >
        Un USER solo puede ver sus propias solicitudes.
        Un ADMIN puede ver cualquiera.
        Si el USER intenta ver una solicitud ajena → 403.
      tags: [Solicitudes - Usuario]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Detalle de la solicitud
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SolicitudResponse'
        '403':
          description: No tenés permiso para ver esta solicitud
        '404':
          description: Solicitud no encontrada

  /solicitudes/{id}/aprobar:
    put:
      summary: Aprobar solicitud (solo ADMIN)
      description: >
        Cambia el estado a APROBADA.
        Solo permitido si el estado actual es PENDIENTE o EN_REVISION.
        Publica evento solicitud.aprobada en RabbitMQ.
      tags: [Solicitudes - Admin]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Solicitud aprobada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SolicitudResponse'
        '409':
          description: >
            Conflicto — la solicitud ya fue resuelta
            (estado APROBADA o RECHAZADA)
        '403':
          description: Solo administradores pueden aprobar
        '404':
          description: Solicitud no encontrada

  /solicitudes/{id}/rechazar:
    put:
      summary: Rechazar solicitud (solo ADMIN)
      description: >
        Cambia el estado a RECHAZADA con motivo obligatorio.
        Solo permitido si el estado actual es PENDIENTE o EN_REVISION.
        Publica evento solicitud.rechazada en RabbitMQ.
      tags: [Solicitudes - Admin]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RechazarSolicitudRequest'
      responses:
        '200':
          description: Solicitud rechazada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SolicitudResponse'
        '409':
          description: Solicitud ya fue resuelta previamente
        '403':
          description: Solo administradores pueden rechazar
        '404':
          description: Solicitud no encontrada

  /solicitudes/admin/todas:
    get:
      summary: Listar todas las solicitudes (solo ADMIN)
      description: Listado completo con filtros por estado, rango de fechas y monto.
      tags: [Solicitudes - Admin]
      parameters:
        - name: estado
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/EstadoSolicitud'
        - name: fechaDesde
          in: query
          required: false
          schema:
            type: string
            format: date
            example: "2025-01-01"
        - name: fechaHasta
          in: query
          required: false
          schema:
            type: string
            format: date
            example: "2025-06-30"
        - name: page
          in: query
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Lista paginada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SolicitudPageResponse'
        '403':
          description: Solo administradores

  /solicitudes/admin/metricas:
    get:
      summary: Métricas agregadas (solo ADMIN)
      tags: [Solicitudes - Admin]
      responses:
        '200':
          description: Totales por estado y monto aprobado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MetricasResponse'
        '403':
          description: Solo administradores